requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
httpx==0.27.0
cloudinary==1.41.0
python-multipart==0.0.12
google-genai>=1.0.0
supabase>=2.0.0
